Urađen dio sa pitanjima i odgovorima

+++ app/src/main/java/ba/etf/rma21/projekat/data/models/Kviz.kt
@@ -12,7 +12,7 @@ data class Kviz(
     val datumRada: Date? = null,
     @SerializedName("trajanje") val trajanje: Int,
     val nazivGrupe: String? = null,
-    val osvojeniBodovi: Float? = null
+    val osvojeniBodovi: Int? = null
 ) {
     var zavrsen: Boolean = false
     var prekinut: Boolean = false
