+++ app/src/main/java/ba/etf/rma21/projekat/data/models/KvizTaken.kt
@@ -6,7 +6,7 @@ import java.util.*
 data class KvizTaken(
     @SerializedName("id") val id: Int,
     @SerializedName("student") val student: String,
-    @SerializedName("osvojeniBodovi") val osvojeniBodovi: Float?,
+    @SerializedName("osvojeniBodovi") val osvojeniBodovi: Int?,
     @SerializedName("datumRada") val datumRada: Date,
     @SerializedName("KvizId") val KvizId: Int
 )
